Example for VpcSourceIp for Resource Policies
This change adds an example for using VpcSourceIp on the API Gateway resource policy to allow customers filtering the traffic to a Private API Gateway API by CIDRs. As VPC Endpoints re-writes the SourceIP of a request, it is necessary to use the VpcSourceIp condition key to obtain the IP of the original caller.