Added: clarification of validation expression use
In order to use an access token with OAuth Scopes for an API method. The token validation expression must be blank/undefined. Otherwise the API will reject the request with a 401 error code due to the Access token not containing an `aud` field. 

This additional information in the documentation is required as this caveat is not documented anywhere and must be derived from the token structure and this `aud` field validation.