Make distinction between API and API endpoint
When reading this page for the first time, I was confused because in
paragraph 2, it says "the resulting API is edge-optimized", but in
paragraph 3, it says "the API is region-specific". Only after reading a
few more docs and articles, I realized that it is only the API endpoint
that is edge-optimized, which the API itself is regional. I have tried
to clarify this by making a distinction between the API and API
endpoint.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.