Fix for UserRoute53Nameservers description
The description in this document seems to suggest the opposite to the description in the serverless application repository (api-gateway-dev-portal — version 4.1.0):

**This doc:**

```UseRoute53Nameservers
Only applicable if you're creating a custom domain name for your developer portal. Specify true to skip creating a Route 53 HostedZone and RecordSet. You'll need to provide your own name server hosting in place of Route 53. Otherwise, leave this field set to false.```

**api-gateway-dev-portal — version 4.1.0:**

```UseRoute53Nameservers
Only applicable if creating a custom domain name for your dev portal. Defaults to false, and you'll need to provide your own nameserver hosting. If set to true, a Route53 HostedZone and RecordSet are created for you.```

I'm proposing to change it to:

```UseRoute53Nameservers
Only applicable if you're creating a custom domain name for your developer portal. If set to true, a Route 53 HostedZone and RecordSet are created for you. Otherwise, leave this field set to false and provide your own name server hosting in place of Route 53.```

Thanks
Eulogio