Troubleshooting addition for Lambda Authorizer
Adding note to troubleshooting tips. 
In the case where a request made does not include the identity source, the api gateway returns a 401, without a hint as to why. There's no authoerizer error because the authorizer is never called. This document, as it should, instructs people to use the Authorization header, however it's not something included when running a quick curl / browser GET test. 
In this case it kind of fails silently, and isnt obvious to figure out. This note would have saved me some time. My lambda didnt need a specific header, it authenticates the request in an unconventional way.